requirements.txt
lxml==4.9.1
tqdm==4.64.0
numpy==1.23.2
torch==1.12.1
pandas==1.4.4
pillow==9.2.0
requests==2.27.1
imagehash==4.2.1
selenium==3.141.0
torchvision==0.13.1
tensorboard==2.10.0
fake-headers==1.0.2
beautifulsoup4==4.11.1
